feat(chat+offers): integrate chat tab in profile, monetary int enforcement, UI polish
- user_profile_screen: add tab 3 (Mensajes) wired to ChatListScreen embedded mode;
  fix _buildMessagesTab placed outside class (compile error)
- app_router: use int.tryParse for askingPrice query param (monetary int rule)
- make_offer_screen: enforce CurrencyInputFormatter on amount field
- property_step2_details_price: enforce CurrencyInputFormatter on price field
- offer_management_screen: counter-offer amount displayed with int formatter
- transaction_timeline_screen: minor title/CTA fixes (already committed states)
- search_provider: deduplicate filter logic; price range as int bounds
- filter_sidebar: price slider values aligned to int monetary rule
- property_details_screen: remove unused color vars (cleanup)
- home_screen: remove unused local vars (cleanup)
- property_form_provider: remove unused dart:io import

+++ frontend/lib/config/router/app_router.dart
@@ -162,7 +162,7 @@ final appRouterProvider = Provider<GoRouter>((ref) {
         builder: (context, state) {
           final propertyId = state.pathParameters['id']!;
           final askingPrice =
-              double.tryParse(state.uri.queryParameters['price'] ?? '0') ?? 0;
+              int.tryParse(state.uri.queryParameters['price'] ?? '0') ?? 0;
           return MakeOfferScreen(propertyId: propertyId, askingPrice: askingPrice);
         },
       ),
